# src/docs/swagger.yaml
basePath: /
definitions:
...
  carwise.ChatInfo:
    description: Chat information
    properties:
      last_message_time:
        type: integer
      user:
        $ref: '#/definitions/carwise.UserInfo'
    type: object
...
  carwise.GetChatsResponse:
    description: Response body for getting chats
    properties:
      chats:
        items:
          $ref: '#/definitions/carwise.ChatInfo'
        type: array
      total:
        type: integer
    type: object
...
  carwise.GetMessagesResponse:
    description: Response body for getting messages
    properties:
      messages:
        items:
          $ref: '#/definitions/carwise.MessageInfo'
        type: array
      total:
        type: integer
    type: object
  carwise.Image:
    description: Image information
    properties:
      created_at:
        description: Timestamp when the image was uploaded
        example: 1646092800
        type: integer
      created_by:
        description: ID of the user who uploaded the image
        example: user_123
        type: string
      id:
        description: Unique identifier of the image
        example: img_123456
        type: string
      path:
        description: Path where the image is stored
        example: /uploads/img123.jpg
        type: string
    type: object
...
  carwise.MessageInfo:
    description: Message information
    properties:
      created_at:
        type: integer
      id:
        type: string
      message:
        type: string
      read:
        type: boolean
      receiver:
        $ref: '#/definitions/carwise.UserInfo'
      sender:
        $ref: '#/definitions/carwise.UserInfo'
    type: object
...
  carwise.PredictImageResponse:
    description: Response containing the prediction results for an image
    properties:
      confidence:
        description: Confidence score of the prediction
        example: 0.95
        type: number
      created_at:
        description: Timestamp of the prediction
        example: 1646092800
        type: integer
      image:
        allOf:
        - $ref: '#/definitions/carwise.Image'
        description: Image information
      prediction:
        description: Whether the image shows car damage
        example: true
        type: boolean
    type: object
  carwise.ProfileResponse:
    description: User profile response
    properties:
      country_code:
        example: "+90"
        type: string
      created_at:
        example: 1646092800
        type: integer
      email:
        example: john.doe@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      image_url:
        example: https://example.com/avatar.jpg
        type: string
      last_name:
        example: Doe
        type: string
      phone_number:
        example: "5551234567"
        type: string
      role:
        example: 1
        type: integer
      status:
        example: 1
        type: integer
    type: object
...
  carwise.SendMessageRequest:
    description: Request body for sending a message
    properties:
      message:
        type: string
    type: object
...
  carwise.UserInfo:
    description: User information
    properties:
      country_code:
        example: "+90"
        type: string
      email:
        example: john.doe@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      last_name:
        example: Doe
        type: string
      phone_number:
        example: "5551234567"
        type: string
    type: object
...
  main.ErrorResponse:
    description: Error response
    properties:
      error:
        example: Invalid request
        type: string
    type: object
  main.SuccessResponse:
    description: Success response
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@carwise.com
    name: API Support
  description: Carwise API documentation with Swagger
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Carwise API
  version: "1.0"
paths:
...
  /chat:
    get:
      consumes:
      - application/json
      description: Get all chats for a user
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Chats retrieved successfully
          schema:
            $ref: '#/definitions/carwise.GetChatsResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get chats
      tags:
      - Message
  /chat/{receiver_id}:
    get:
      consumes:
      - application/json
      description: Get messages between two users
      parameters:
      - description: Receiver ID
        in: path
        name: receiver_id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Messages retrieved successfully
          schema:
            $ref: '#/definitions/carwise.GetMessagesResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get messages
      tags:
      - Message
    post:
      consumes:
      - application/json
      description: Send a message to a user
      parameters:
      - description: Receiver ID
        in: path
        name: receiver_id
        required: true
        type: string
      - description: Message details
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/carwise.SendMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Message sent successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send a message
      tags:
      - Message
...
  /profile:
    get:
      description: Get the profile information of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User profile information
          schema:
            $ref: '#/definitions/carwise.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Profile
    put:
      consumes:
      - multipart/form-data
      description: Update the profile information of the authenticated user
      parameters:
      - description: First name
        in: formData
        name: first_name
        required: true
        type: string
      - description: Last name
        in: formData
        name: last_name
        required: true
        type: string
      - description: Country code
        in: formData
        name: country_code
        required: true
        type: string
      - description: Phone number
        in: formData
        name: phone_number
        required: true
        type: string
      - description: Profile picture
        in: formData
        name: avatar
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Edit user profile
      tags:
      - Profile
  /upload/:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image file (max 5MB, jpeg/png/gif)
      parameters:
      - description: Image file to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Successfully uploaded image
          schema:
            $ref: '#/definitions/carwise.Image'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload an image
      tags:
      - Upload
  /upload/{id}:
    delete:
      description: Delete an image by its ID
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted image
          schema:
            $ref: '#/definitions/main.SuccessResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an image
      tags:
      - Upload
  /upload/{id}/predict:
    get:
      description: Use AI to predict if an image shows car damage
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Prediction results
          schema:
            $ref: '#/definitions/carwise.PredictImageResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Predict car damage from image
      tags:
      - Upload
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"